fix more warnings emerged by Travis CI

+++ Source/m_fixed.h
@@ -64,7 +64,7 @@ __inline__ static fixed_t FixedMul(fixed_t a, fixed_t b)
 
 __inline__ static fixed_t FixedDiv(fixed_t a, fixed_t b)
 {
-  return (abs(a)>>14) >= abs(b) ? ((a^b)>>31) ^ D_MAXINT :
+  return (abs(a)>>14) >= abs(b) ? ((unsigned)(a^b)>>31) ^ D_MAXINT :
     (fixed_t)(((Long64) a << FRACBITS) / b);
 }
